fix(exports): Move exports to Components/index.ts
