--- winkduino-app/Components/StatusBanner.tsx
@@ -1,15 +1,0 @@
-import { View } from "react-native";
-
-
-export function StatusBanner() {
-  return (
-    <View
-      style={{
-        position: "absolute",
-
-      }}
-    >
-
-    </View>
-  )
-}
