+++ winkduino-app/Pages/Home/CustomCommands.tsx
@@ -5,7 +5,7 @@ import { useNavigation, useRoute } from "@react-navigation/native";
 import { useEffect, useState } from "react";
 import { useBLE } from "../../hooks/useBLE";
 import { ActivityIndicator } from "react-native";
-import { HeaderWithBackButton } from "../../Components/HeaderWithBackButton";
+import { HeaderWithBackButton } from "../../Components";
 
 export function CustomCommand() {
   const { theme, colorTheme } = useColorTheme();
